organized files & update saved_route.c
